pyproject.toml
[tool.poetry]
name = "5a"
version = "0.0.1"
description = "A DSPy-based project"
authors = ["Arthur <arthur@0xzap.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
dspy-ai = "^2.4.0"
python-dotenv = "^1.0.0"
mem0ai = "^0.1.114"
pinecone = "^7.3.0"
pinecone-text = "^0.10.0"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true 